remove nixpkgs pinning
For a 140 lines script that only depends on 3 dependencies it seems rather an
overkill to pin nipxkgs as it will result in a huge downloads and more
maintainance burdens to keep .nixpkgs-version.json up to date. All libraries
used here are mature and not likely to break across different versions. Also if
they would break it would be better to catch this early in the repository by
having an up to date nixpkgs revision rather than a pinned one.

Signed-off-by: Jörg Thalheim <joerg@thalheim.io>